requirements.txt
django
gunicorn
redis
flake8
pytest
pytest-cov
coverage
mysqlclient
djangorestframework
djangorestframework-simplejwt
markdown
django-filter
django-oauth-toolkit
django-cors-headers